Adding a test harness, fixing some bugs and adding things

+++ app/src/main/java/com/ewisnor/randomur/imgur/model/GalleryAlbum.java
@@ -7,7 +7,7 @@ public class GalleryAlbum {
     private String id;
     private String title;
     private String description;
-    private Integer dateTime;
+    private Long dateTime;
     private String cover;
     private Integer coverWidth;
     private Integer coverHeight;
@@ -40,7 +40,7 @@ public class GalleryAlbum {
         return description;
     }
 
-    public Integer getDateTime() {
+    public Long getDateTime() {
         return dateTime;
     }
